fix a circular reference by splitting into several files
