+++ source/views/building-hours/lib/chapel.js
@@ -3,7 +3,7 @@ import moment from 'moment-timezone'
 import type {SingleBuildingScheduleType} from '../types'
 
 import {getDayOfWeek} from './get-day-of-week'
-import {isScheduleReallyOpenAtMoment} from './is-schedule-open'
+import {isScheduleReallyOpenAtMoment} from './is-schedule-really-open'
 import {formatBuildingTimes} from './format-times'
 import {parseHours} from './parse-hours'
