+++ source/views/building-hours/lib/is-schedule-open.js
@@ -2,16 +2,8 @@
 import moment from 'moment-timezone'
 import type {SingleBuildingScheduleType} from '../types'
 
-import {parseHours} from './parse-hours'
 import {isChapelTime} from './chapel'
-
-export function isScheduleReallyOpenAtMoment(
-	schedule: SingleBuildingScheduleType,
-	m: moment,
-): boolean {
-	let {open, close} = parseHours(schedule, m)
-	return m.isBetween(open, close, 'minute', '[)')
-}
+import {isScheduleReallyOpenAtMoment} from './is-schedule-really-open'
 
 export function isScheduleOpenAtMoment(
 	schedule: SingleBuildingScheduleType,
